+++ NEWS.md
@@ -26,6 +26,9 @@
   - New `append_parquet()` function to append a data frame to an
     existing Parquet file.
 
+  - New `col_select` argument for `read_parquet()` to read a subset of
+    columns from a Parquet file.
+
   - `write_parquet()` can now write multiple row groups. By default it puts
     at most 10 million rows into a single row group. You can choose the
     row groups manually with the `row_groups` argument.